+++ tests/rest/test_products.py
@@ -3,7 +3,8 @@ import unittest
 from requests_mock import Mocker
 
 from coinbase.rest import RESTClient
-from tests.constants import TEST_API_KEY, TEST_API_SECRET
+
+from ..constants import TEST_API_KEY, TEST_API_SECRET
 
 
 class ProductsTest(unittest.TestCase):